requirements.txt
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
cryptography>=41.0.0
python-dotenv>=1.0.0
openai>=1.0.0
scikit-learn>=1.3.0
nltk>=3.8.0
transformers>=4.30.0
torch>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
alembic>=1.11.0
asyncio-mqtt>=0.13.0
keyring>=24.0.0
pandas>=1.5.0